making consumer_update callback able to call async methods
