+++ examples/single_active_consumer/single_active_consumer.rs
@@ -4,12 +4,13 @@ use rabbitmq_stream_client::types::{
     ByteCapacity, OffsetSpecification, ResponseCode,
 };
 
+
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     use rabbitmq_stream_client::Environment;
     let environment = Environment::builder().build().await?;
     let message_count = 1000000;
-    let stream = "hello-rust-stream";
+    let stream = "hello-rust-super-stream-2";
 
     let create_response = environment
         .stream_creator()
@@ -40,15 +41,31 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .offset(OffsetSpecification::First)
         .enable_single_active_consumer(true)
         .client_provided_name("my super stream consumer for hello rust")
-        .consumer_update(move |active, message_context| {
-            println!("single active consumer: is active: {} on stream {}", active, message_context.get_stream());
-            OffsetSpecification::First
+        .consumer_update(move |active, message_context|  async move {
+            let name = message_context.name();
+            let stream = message_context.stream();
+            let client = message_context.client();
+
+            println!(
+                "single active consumer: is active: {} on stream: {} with consumer_name: {}",
+                active, stream, name
+            );
+            let stored_offset = client.query_offset(name, stream.as_str()).await;
+
+            if let Err(e) = stored_offset {
+                return OffsetSpecification::First;
+            }
+
+            let stored_offset_u = stored_offset.unwrap();
+            println!("restarting from stored_offset:  {}", stored_offset_u);
+            OffsetSpecification::Offset(stored_offset_u)
+
         })
         .build(stream)
         .await
         .unwrap();
 
-    for _ in 0..message_count {
+    for i in 0..message_count {
         let delivery = consumer.next().await.unwrap();
         {
             let delivery = delivery.unwrap();
@@ -62,6 +79,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 delivery.stream(),
                 delivery.offset()
             );
+
+            //store an offset
+            if i == 10000  {
+                let _ = consumer
+                    .store_offset(i)
+                    .await
+                    .unwrap_or_else(|e| println!("Err: {}", e));
+            }
         }
     }
